Revert "increment bsddialog_total_progview"
This reverts commit ad49e37e98701123f0133213038f8cfcbfe69840

+++ lib/barbox.c
@@ -45,7 +45,7 @@
 
 bool bsddialog_interruptprogview;
 bool bsddialog_abortprogview;
-long long int bsddialog_total_progview;
+int  bsddialog_total_progview;
 
 static const char states[12][14] = {
 	"  Succeeded  ", /* -1  */
